fix(platform): icon-tab-bar types & interfaces
