+++ libs/platform/src/lib/icon-tab-bar/components/icon-tab-bar-text-type/icon-tab-bar-text-type.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectorRef, Component, EventEmitter, Input, NgZone, Output } from '@angular/core';
 import { IconTabBarBase } from '../icon-tab-bar-base.class';
-import { IconTabBarItem } from '../../types';
+import { IconTabBarItem } from '../../interfaces/icon-tab-bar-item.interface';
 import { UNIQUE_KEY_SEPARATOR } from '../../constants';
 import { FdDnDEvent } from '../../directives/dnd/icon-bar-dnd-container.directive';
