+++ libs/platform/src/lib/icon-tab-bar/icon-tab-bar.component.ts
@@ -10,7 +10,9 @@ import {
     Output,
     ViewEncapsulation
 } from '@angular/core';
-import { IconTabBarBackground, IconTabBarItem, IconTabBarSize, TabConfig, TabDestinyMode, TabType } from './types';
+import { IconTabBarItem } from './interfaces/icon-tab-bar-item.interface'
+import { TabConfig } from './interfaces/tab-config.interface';
+import { IconTabBarBackground, IconTabBarSize, TabDestinyMode, TabType } from './types';
 import { ContentDensityService, RtlService } from '@fundamental-ngx/core/utils';
 import { IconFont, } from '@fundamental-ngx/core/icon';
 import { distinctUntilChanged, takeUntil } from 'rxjs/operators';
