+++ libs/platform/src/lib/icon-tab-bar/interfaces/icon-tab-bar-item.interface.ts
@@ -1,17 +1,6 @@
-import { SemanticColor } from './index';
-
-// This is config that user should provide to input
-export interface TabConfig {
-    icon?: string;
-    label?: string;
-    color?: SemanticColor;
-    counter?: number;
-    active?: boolean; // selected tab
-    badge?: boolean; // should show red circle in top-right corner of tab
-    subItems?: TabConfig[]; // Nested tabs
-}
-
 // This is extended TabConfig with special fields
+import { TabConfig } from './tab-config.interface';
+
 export interface IconTabBarItem extends TabConfig {
     index: number; // order in the array
     /*
