--- libs/platform/src/lib/icon-tab-bar/types/index.ts
@@ -1,2 +1,0 @@
-export * from './common.types';
-export * from './icon-tab-bar-items.interface';
